Implements a Python client (#116)
* moving lua files to their own directory
* Adding some comments
* Adding an example
* Initial commit of python torchcraft. Replayer loading/saving works
* Work on constants
* Added the rest of the python client
* Making flags properties
* Adds more documentation
* Making the setup script a bit simpler
