+++ lua/_env.lua
@@ -1,0 +1,2 @@
+local tc = {}
+return tc
