+++ lua/replayer.lua
@@ -1,0 +1,1 @@
+return require('torchcraft.tc_lib').replayer
