+++ py/requirements.txt
@@ -1,0 +1,1 @@
+pybind11>=2.1.0
