--- replayer.lua
@@ -1,1 +1,0 @@
-return require('torchcraft.tc_lib').replayer
