+++ torchcraft-1.2-1.rockspec
@@ -22,6 +22,14 @@ build = {
     build_command = [[
     cmake -E make_directory build && cd build && cmake .. -DCMAKE_BUILD_TYPE=Release -DLUALIB=$(LUALIB) -DLUA_INCDIR="$(LUA_INCDIR)" -DLUA_LIBDIR="$(LUA_LIBDIR)" -DLUADIR="$(LUADIR)" -DLIBDIR="$(LIBDIR)" -DCMAKE_INSTALL_PREFIX="$(PREFIX)" && $(MAKE)
     ]],
-    install_command = "cd build && $(MAKE) install"
+    install_command = "cd build && $(MAKE) install",
+    install = {
+      lua = { 
+        ["torchcraft._env"] = "lua/_env.lua",
+        ["torchcraft.init"] = "lua/init.lua",
+        ["torchcraft.replayer"] = "lua/replayer.lua",
+        ["torchcraft.utils"] = "lua/utils.lua"
+      }
+    }
 }
